pyproject.toml
[project]
name = "loftusa-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.75.0",
    "bitsandbytes>=0.49.1",
    "cerebras-cloud-sdk>=1.59.0",
    "datasets>=4.4.1",
    "dotenv>=0.9.9",
    "fastapi>=0.121.2",
    "groq>=0.37.1",
    "ipykernel>=7.1.0",
    "litellm>=1.80.13",
    "matplotlib>=3.8.0",
    "openai>=2.8.0",
    "openai-agents>=0.6.2",
    "peft>=0.18.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "transformers[torch]>=4.57.3",
    "trl[peft]>=0.26.2",
    "twilio>=9.8.6",
    "typeguard>=4.4.4",
    "uvicorn>=0.38.0",
    "rich>=14.2.0",
    "inspect-ai>=0.3.164",
    "faiss-cpu>=1.7.0",
    "pypdf2>=3.0.0",
]
